+++ Basil/include/Core/Basil.h
@@ -30,3 +30,4 @@
 #include "Scene/Component.h"
 #include "Scene/Entity.h"
 #include "Scene/Scene.h"
+#include "Scene/ScriptableEntity.h"