+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Change log for the Basil Game Engine project.
 * Camera component.
 * SceneCamera class.
 * Camera resizing support.
+* Initial native scripting support.
 
 ### Changed
 * Changed `BIND_EVENT` in `Core.h` from `std::bind` to a lambda.